extract Context interface from implementation
Geebus.  Not like it's the foundation of everything...
