+++ choam/src/main/java/com/salesforce/apollo/choam/ViewAssembly.java
@@ -55,6 +55,7 @@ import com.salesforce.apollo.ethereal.Ethereal.Controller;
 import com.salesforce.apollo.ethereal.Ethereal.PreBlock;
 import com.salesforce.apollo.ethereal.memberships.ContextGossiper;
 import com.salesforce.apollo.membership.Context;
+import com.salesforce.apollo.membership.ContextImpl;
 import com.salesforce.apollo.membership.Member;
 
 import io.grpc.Status;
@@ -108,8 +109,8 @@ public class ViewAssembly implements Reconfiguration {
                                                 new ArrayList<>(nextAssembly.values()), comms);
         // Create a new context for reconfiguration
         final Digest reconPrefixed = view.context().getId().xor(nextViewId);
-        Context<Member> reContext = new Context<Member>(reconPrefixed, view.context().getProbabilityByzantine(),
-                                                        view.context().memberCount(), view.context().getBias());
+        Context<Member> reContext = new ContextImpl<Member>(reconPrefixed, view.context().getProbabilityByzantine(),
+                                                            view.context().memberCount(), view.context().getBias());
         reContext.activate(view.context().activeMembers());
 
         final Fsm<Reconfiguration, Transitions> fsm = Fsm.construct(this, Transitions.class, getStartState(), true);
