+++ choam/src/test/java/com/salesforce/apollo/choam/SessionTest.java
@@ -37,6 +37,7 @@ import com.salesforce.apollo.choam.support.InvalidTransaction;
 import com.salesforce.apollo.choam.support.SubmittedTransaction;
 import com.salesforce.apollo.crypto.DigestAlgorithm;
 import com.salesforce.apollo.membership.Context;
+import com.salesforce.apollo.membership.ContextImpl;
 import com.salesforce.apollo.membership.Member;
 import com.salesforce.apollo.membership.impl.SigningMemberImpl;
 import com.salesforce.apollo.utils.Utils;
@@ -58,7 +59,7 @@ public class SessionTest {
     @Test
     public void func() throws Exception {
         ScheduledExecutorService exec = Executors.newSingleThreadScheduledExecutor();
-        Context<Member> context = new Context<>(DigestAlgorithm.DEFAULT.getOrigin(), 9);
+        Context<Member> context = new ContextImpl<>(DigestAlgorithm.DEFAULT.getOrigin(), 9);
         Parameters params = Parameters.newBuilder()
                                       .build(RuntimeParameters.newBuilder()
                                                               .setContext(context)
@@ -94,7 +95,7 @@ public class SessionTest {
     @Test
     public void scalingTest() throws Exception {
         ScheduledExecutorService exec = Executors.newSingleThreadScheduledExecutor();
-        Context<Member> context = new Context<>(DigestAlgorithm.DEFAULT.getOrigin(), 9);
+        Context<Member> context = new ContextImpl<>(DigestAlgorithm.DEFAULT.getOrigin(), 9);
         Parameters params = Parameters.newBuilder()
                                       .setTxnPermits(100_000_000)
                                       .build(RuntimeParameters.newBuilder()
