--- model/src/main/java/com/salesforce/apollo/model/MemberState.java
@@ -1,19 +1,0 @@
-/*
- * Copyright (c) 2021, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-package com.salesforce.apollo.model;
-
-/**
- * @author hal.hildebrand
- *
- */
-public enum MemberState {
-    OFFLINE, ACTIVE, JOINING, LEAVING;
-
-    public String toColumn() {
-        return name().toLowerCase();
-    }
-}
